requirements.txt
fpdf==1.7.2
python_docx==1.1.2
streamlit==1.40.0
